+++ vice/src/c64/cart/retroreplay.c
@@ -261,7 +261,7 @@ void REGPARM2 retroreplay_io1_store(WORD addr, BYTE value)
                         value = 0;
                     }
                 }
-                cartridge_config_changed(0, (value & 3) | (rr_bank << CMODE_BANK_SHIFT), mode);
+                cartridge_config_changed(0, (BYTE)((value & 3) | (rr_bank << CMODE_BANK_SHIFT)), mode);
 
                 if (value & 4) {
                     rr_active = 0;
